restructure FSM mechanic per PR review feedback
- split Runtime/Core/StateMachine into Scripts/ and Script_Explainers/
- add per-script explainer for every runtime .cs (mandatory)
- move ReadOnlyAttribute under StateMachine/Scripts/ (added with this mechanic)
- replace loose Samples~/StateMachineSample/Assets folder with StateMachineSample.zip
- add zipped tutorial video alongside the existing mp4
- remove aggregate Samples ScriptExplainer.txt now that per-script explainers exist
- update README mechanic entry paths and add co-authors
