--- Runtime/Core/ReadOnlyAttribute.cs
@@ -1,7 +1,0 @@
-using UnityEngine;
-
-namespace GameplayMechanicsUMFOSS.Core
-{
-    /// <summary>Makes a serialized field visible in the Inspector but not editable.</summary>
-    public class ReadOnlyAttribute : PropertyAttribute { }
-}
