--- Runtime/Core/ReadOnlyAttribute.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 39900891da5e448eb2b9c8cfdda7b89a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
