--- Runtime/Core/StateMachine/IState_UMFOSS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 86372b8b9f6f46ab89df9e5633fb9935
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
