+++ Runtime/Core/StateMachine/Scripts/ReadOnlyAttribute.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+namespace GameplayMechanicsUMFOSS.Core
+{
+    /// <summary>Makes a serialized field visible in the Inspector but not editable.</summary>
+    public class ReadOnlyAttribute : PropertyAttribute { }
+}
