+++ Runtime/Core/StateMachine/Scripts/StateMachineDebugger_UMFOSS.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15a730b100b9469bbf14306fa381959e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
