+++ Runtime/Core/StateMachine/Scripts/StateMachine_UMFOSS.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4b53607989e04a7f93bd424d8ec1bc16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
