--- Runtime/Core/StateMachine/StateHistory_UMFOSS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eb37b9d71a894e73a3ac896bca6964f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
