--- Runtime/Core/StateMachine/StateMachineEvents_UMFOSS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: addd240616f14587b02e0d2ff9bbe310
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
