--- Runtime/Core/StateMachine/StateTransition_UMFOSS.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 696d7b5c9091467b8576e5160bc57358
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
