--- Samples~/StateMachineSample/Assets/Scenes/DemoScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8982c9c731ab446a586726f81a78ff59
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
