--- Samples~/StateMachineSample/Assets/Scripts/GroundDetector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7b7ec72e529c4121952d0719852f9100
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
